+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 — 2022-01-02
+
+### Added
+
+- `purge` slash and context commands! 🌟
+  - Useful for when creating a new thread isn't necessary, and you just want to clean up spam.
+  - Use `/purge <messages>` (with a maximum of 100) to delete the specified amount of messages, starting with the most recent.
+  - Right-click a message within 100 recent messages in a channel to delete all the messages from there onward.
+  - Only available for server members with the `Manage Messages` permission.
+
 ## 0.2.1 — 2021-11-27
 
 ### Changed